feat: udpated the Ds Input component
