+++ packages/@sparrow-common/src/features/create-team/components/team-name/TeamName.svelte
@@ -13,7 +13,7 @@
    * Constants
    */
   import { NAME_CONFIG } from "../../constants";
-  import { NewInput } from "@sparrow/library/forms";
+  // import { NewInput } from "@sparrow/library/forms";
 
   /**
    * Exports
@@ -55,32 +55,7 @@
   <!-- 
     -- Input 
   -->
-  <!-- <Input
-    bind:value={teamForm.name.value}
-    on:blur={() => {
-      teamForm.name.isTouched = true;
-      teamForm.name.value = teamForm.name.value.trim(); // Trim the value on blur
-    }}
-    height={"36px"}
-    id={inputId}
-    placeholder={NAME_CONFIG.PLACEHOLDER}
-    class="text-fs-14 bg-tertiary-300 fw-normal px-2 border-radius-4"
-    style="outline:none;"
-    defaultBorderColor={!teamForm.name.value && teamForm.name.isTouched
-      ? "var(--border-danger-200)"
-      : "transparent"}
-    hoveredBorderColor={!teamForm.name.value && teamForm.name.isTouched
-      ? "var(--border-danger-200)"
-      : "transparent"}
-    focusedBorderColor={!teamForm.name.value && teamForm.name.isTouched
-      ? "var(--border-danger-200)"
-      : "transparent"}
-    isEditIconRequired={false}
-    type={"text"}
-    maxlength={NAME_CONFIG.MAX_TEXT_SIZE}
-    placeholderColor={"var(--text-secondary-200)"}
-  /> -->
-  <NewInput
+  <Input
     bind:value={teamForm.name.value}
     on:blur={() => {
       teamForm.name.isTouched = true;
@@ -99,16 +74,6 @@
     style="outline:none;"
     class="text-fs-14 fw-normal py-2 px-1  border-radius-4"
   />
-
-  <!-- defaultBorderColor={!teamForm.name.value && teamForm.name.isTouched
-    ? "1px solid var(--border-ds-danger-300) !important"
-    : "transparent"}
-  hoveredBorderColor={!teamForm.name.value && teamForm.name.isTouched
-    ? "1px solid var(--border-ds-danger-300) !important"
-    : "transparent"}
-  focusedBorderColor={!teamForm.name.value && teamForm.name.isTouched
-    ? "2px solid var(--border-ds-danger-300) !important"
-    : "transparent"} -->
 
   <!-- 
     -- Error Messages
