+++ packages/@sparrow-library/src/forms/Input/Input.svelte
@@ -2,147 +2,105 @@
   import { SearchIcon } from "@sparrow/library/assets";
   import { PencilIcon } from "@sparrow/library/icons";
   import { createEventDispatcher } from "svelte";
-
-  /**
-   * input type
-   */
-  export let type: "text" | "password" | "search" = "text";
-  /**
-   * placeholder - dummy text
-   */
-  export let placeholder = "placeholder";
-  export let placeholderColor = "gray";
-  /**
-   * height
-   */
-  export let height = "30px";
-  /**
-   * width
-   */
-  export let width = "auto";
-
-  /**
-   * identifies input is disabled or not
-   */
-  export let disabled = false;
-  /**
-   * input class
-   */
   let componentClass = "";
   export { componentClass as class };
-
-  /**
-   * input style
-   */
   let componentStyle = "";
   export { componentStyle as style };
-  /**
-   * input value
-   */
+
+  export let type: "text" | "password" | "search" = "text";
+  export let placeholder = "placeholder";
+  export let placeholderColor = "gray";
+  export let height = "36px";
+  export let width = "auto";
+  export let disabled = false;
   export let value = "";
-
-  /**
-   * border color
-   */
   export let defaultBorderColor = "transparent";
-  export let hoveredBorderColor = "red";
-  export let focusedBorderColor = "green";
-
+  export let typingBorderColor = "1px solid var(--border-ds-primary-300)";
+  export let hoveredBorderColor = "1px solid var(--border-ds-neutral-300)";
+  export let focusedBorderColor = "2px solid var(--border-ds-primary-300)";
+  export let typedBorderColor = "";
+  export let defaultBgColor = "var(--bg-ds-surface-400)";
+  export let disabledBgColor = "var(--bg-ds-surface-600)";
   export let isEditIconRequired = true;
-
   export let maxlength = 500;
-
   export let searchIconColor = "var(--defaultcolor)";
   export let iconSize = "14px";
-
-  /**
-   * Unique id for input
-   */
   export let id = "";
 
-  /**
-   * input states
-   */
   let isHovered = false;
   let isFocused = false;
+  let isTyping = false;
+  let hasInput = false;
   const dispatch = createEventDispatcher();
 
-  const extractBorderHighlight = (
-    _isHovered: boolean,
-    _isFocused: boolean,
-    _defaultBorderColor: string,
-    _hoveredBorderColor: string,
-    _focusedBorderColor: string,
-  ) => {
-    if (_isFocused) {
-      return _focusedBorderColor;
-    } else if (_isHovered) {
-      if (disabled) return _defaultBorderColor;
-      return _hoveredBorderColor;
-    } else {
-      if (disabled) return _defaultBorderColor;
-      return _defaultBorderColor;
-    }
-  };
+  $: hasInput = value.length > 0;
+  $: borderColor = isTyping
+    ? typingBorderColor
+    : isFocused
+      ? focusedBorderColor
+      : isHovered
+        ? hoveredBorderColor
+        : hasInput
+          ? typedBorderColor
+          : defaultBorderColor;
+  $: backgroundColor = disabled ? disabledBgColor : defaultBgColor;
 
-  /**
-   * blur input on Enter key press
-   * @param event - keyboard event
-   */
   const onKeyPress = (event: KeyboardEvent) => {
     if (event.key === "Enter") {
       const inputField = document.getElementById(id) as HTMLInputElement;
-      inputField.blur();
+      inputField?.blur();
+    }
+  };
+
+  const handleInputChange = (event: Event) => {
+    const target = event.target as HTMLInputElement;
+    value = target.value;
+    dispatch("input", value);
+
+    if (typingBorderColor !== "transparent") {
+      isTyping = true;
+      setTimeout(() => {
+        if (value === target.value) {
+          isTyping = false;
+        }
+      }, 2000);
     }
   };
 </script>
 
 <div
-  class="position-relative"
-  style="height:{height}; width: {width}; !important"
-  on:mouseenter={() => {
-    isHovered = true;
-  }}
-  on:mouseleave={() => {
-    isHovered = false;
-  }}
+  class="position-relative {componentClass}"
+  style={`height: ${height}; width: ${width}; background-color: ${backgroundColor}; border: ${borderColor}; border-radius:4px`}
+  on:mouseenter={() => (isHovered = true)}
+  on:mouseleave={() => (isHovered = false)}
 >
   <input
-    {id}
-    on:focus={() => {
-      isFocused = true;
-    }}
-    on:blur={() => {
-      isFocused = false;
-      dispatch("blur", event?.target?.value);
-    }}
     {value}
-    on:input={(event) => {
-      value = event?.target?.value;
-      dispatch("input", event?.target?.value);
-    }}
-    on:keydown={onKeyPress}
+    {id}
     {type}
-    {maxlength}
-    class=" w-100 {componentClass}"
     {placeholder}
-    style=" {componentStyle} height: 100%; {type === 'search'
-      ? `padding-left:${height} !important;`
-      : ''} {type === 'text' && isEditIconRequired && isHovered
-      ? 'padding-right:35px !important;'
-      : ''} border-color:{extractBorderHighlight(
-      isHovered,
-      isFocused,
-      defaultBorderColor,
-      hoveredBorderColor,
-      focusedBorderColor,
-    )}; --placeholder-color: {placeholderColor};"
+    {maxlength}
     {disabled}
+    class="w-100"
+    style="{componentStyle} height: 100%;
+      {type === 'search' ? `padding-left: ${height} !important;` : ''}
+      {type === 'text' && isEditIconRequired && isHovered
+      ? 'padding-right:35px !important;'
+      : ''}
+      --placeholder-color: {placeholderColor}; background-color: ${backgroundColor};"
+    on:focus={() => (isFocused = true)}
+    on:blur={(event) => {
+      isFocused = false;
+      dispatch("blur", event.target?.value);
+    }}
+    on:input={handleInputChange}
+    on:keydown={onKeyPress}
   />
+
   {#if type === "search"}
     <span
-      class="position-absolute d-flex align-items-center justify-content-center m-0 p-0"
-      style="top: 0; left: 0; bottom: 0; width: {height}; "
+      class="position-absolute d-flex align-items-center justify-content-center"
+      style="top: 0; left: 0; bottom: 0; width: {height};"
     >
       <span class="SearchIconClass" style="margin-top:1px;">
         <SearchIcon
@@ -153,20 +111,23 @@
       </span>
     </span>
   {/if}
+
   {#if type === "text" && isHovered && isEditIconRequired && !disabled}
-    <span class="position-absolute" style="top:2px; right: 10px">
-      <PencilIcon height={iconSize} width={iconSize} color={"white"} />
+    <span
+      class="position-absolute"
+      style="top: {height === '36px' ? '4px' : '2px'}; right: 10px;"
+    >
+      <PencilIcon height={iconSize} width={iconSize} color="white" />
     </span>
   {/if}
 </div>
 
 <style>
-  .SearchIconClass {
-    display: flex;
-  }
   input {
-    caret-color: var(--border-primary-300);
+    caret-color: var(--border-ds-primary-300);
     border: 1px solid transparent;
+    min-width: 240px;
+    max-width: 540px;
   }
   input::placeholder {
     color: var(--placeholder-color);
