+++ packages/@sparrow-library/src/forms/index.ts
@@ -5,7 +5,5 @@ export { default as Textarea } from "./textarea/Textarea.svelte";
 export { default as Editor } from "./editor/Editor.svelte";
 export { default as MultiSelect } from "./multi-select/MultiSelect.svelte";
 export { default as TextEditor } from "./text-editor/TextEditor.svelte";
-export { default as Switch } from "./Switch/Switch.svelte";
-export { default as NewInput } from "./newInput/NewInput.svelte";
-export { default as NewTextArea } from "./newtext-area/NewTextArea.svelte";
+// export { default as Switch } from "./Switch/Switch.svelte";
 export { default as Search } from "./search/Search.svelte";
